requirements.txt
pandas
requests
pytz
python-telegram-bot==13.15
mplfinance
apscheduler